requirements.txt
# Copyright (C) 2018, 2023 Mitsubishi Electric Research Laboratories (MERL)
#
# SPDX-License-Identifier: AGPL-3.0-or-later

gdown>=3.4.6
numpy>=1.11.0
matplotlib>=1.5.1
h5py==2.6.0
glog>=0.3.1
Pillow>=1.1.7
scipy>=0.17.0
scikit_learn>=0.19.1
scikit_image>=0.9.3
pydot>=1.1.0
